feat: partition CSV + named-partition ops, backup/restore, 460800 default
Adds the partition-aware workflow that's the highest-value addition for
LLM/CI use: an agent can now say "flash this to partition ota_0" instead
of computing offsets, and the table itself can either come from a
partitions.csv or be read back from the chip's flash at 0x8000.

New module src/partition.rs:
  - CSV parser (IDF dialect: K/M size suffixes, auto-allocated offsets
    with 64K alignment for app partitions and 4K for data, comments,
    encrypted/readonly flags)
  - Binary parser for the 32-byte 0xAA50-magic records at flash 0x8000;
    verifies the MD5 sidecar record when present; stops at the 0xFF
    end-of-table marker
  - Type / subtype enumeration tracking esp_partition.h
  - Validation: duplicate-name and overlap checks
  - Unit tests cover CSV with auto-allocation, binary round-trip, and
    overlap detection

New subcommands:
  - partitions             list the table (CSV or read from chip)
  - write-partition --name resolve name → write file
  - read-partition --name  resolve name → dump to file
  - erase-partition --name resolve name → erase region
  - backup                 dump entire flash; auto-size from JEDEC RDID
                           capacity byte unless --size given
  - restore                write a backup file from offset 0x0

NDJSON event additions: partition_table_loaded (with source: csv:path
or flash:0xN+0xM), partition_resolved (name+type+subtype+offset+size),
backup_begin/done, restore_begin/done, baud_upgrade.

Default baud is now 460800. Initial sync still happens at 115200 (the
ROM bootloader's safe rate); after the stub is up, we issue change_baud
to upgrade — matching upstream esptool's behavior. Stub's second_arg in
the change_baud command is the previous baud (so it can recompute the
UART divider), or 0 when still on ROM.

Build / clippy / tests clean. 17 tests pass (4 new partition tests).

+++ src/chip.rs
@@ -95,8 +95,8 @@ pub fn esp32p4_stub_selector(
     // EFUSE BLOCK1 word 2 holds the revision bits; mirrored from upstream
     // ESP32P4ROM.get_major_chip_version() / get_minor_chip_version().
     let word = conn.read_reg(chip.efuse_block1_addr + 4 * 2)?;
-    let minor = (word & 0x0F) as u32;
-    let major = ((((word >> 23) & 1) << 2) | ((word >> 4) & 0x03)) as u32;
+    let minor = word & 0x0F;
+    let major = (((word >> 23) & 1) << 2) | ((word >> 4) & 0x03);
     let revision = major * 100 + minor;
     Ok(if revision < 300 {
         "esp32p4-rev1"
